Tests: Test files
The ascii test function test for any errors in the whole code. The tests written are also for some functions.

TestCheckIllegalchar checks for anyillegal characters in a given string not in the ascii range of 32 - 126.
TestHandleNewLine check whether the given string has newline escape characters ,tab,return carriage.
TestStringContain checks whether a string has "o" characters as used in the thinkertoy.If it doest the file will also be split using carriage return.
TestExecCommand if the output art of the whole program matches the one we require. This is done by matching the strings stored in the file and the ones generated within the test program.